Change TG recovery install script filename [1/2]
This is part 1/2 to maintain backwards compatibility with CWM's
verify_root_and_recovery() function. CWM checks if install-recovery.sh
exists and has an executable flag set, then offers to disable the script
for the user. CM now controls this with the persist.sys.recovery_update
property which is configurable via settings, so we don't need to
double-check this.

This changes the name of the recovery install script to
install-tg-recovery.sh.

Change-Id: Ie08d428cfd092a1de10973ba5ffc326ccdb1a369